Remove/move arithmetic expression test and expression type checking test. Remove unnecessary Sqrt type rule.
